pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msgraph-beta-sdk"
version = "1.0.0a3"
authors = [{name = "Microsoft", email = "graphtooling+python@microsoft.com"}]
description = "The Microsoft Graph Beta Python SDK"
dependencies = [
    "azure-identity >=1.10.0",
    "microsoft-kiota-abstractions >=0.2.2",
    "microsoft-kiota-authentication-azure >=0.2.0",
    "microsoft-kiota-serialization-json >=0.2.1",
    "microsoft-kiota-serialization-text >=0.2.0",
    "msgraph_core >=1.0.0a0"

]
requires-python = ">=3.6"
license = {file = "LICENSE"}
readme = "README.md"
keywords = ["msgraph", "openAPI", "Microsoft", "Graph", "beta"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
dev = ["yapf", "bumpver", "isort", "pylint", "pytest", "mypy"]

[project.urls]
homepage = "https://github.com/microsoftgraph/msgraph-beta-sdk-python#readme"
repository = "https://github.com/microsoftgraph/msgraph-beta-sdk-python"
documentation = "https://github.com/microsoftgraph/msgraph-beta-sdk-python/docs"

[tool.mypy]
warn_unused_configs = true
files = "msgraph"
ignore_missing_imports = true

[tool.yapf]
based_on_style = "pep8"
dedent_closing_brackets = true
each_dict_entry_on_separate_line = true
column_limit = 100

[tool.isort]
profile = "hug"

[tool.pytest.ini_options]
pythonpath = [
  "msgraph"
]

[tool.bumpver]
current_version = "1.0.0a3"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"msgraph/_version.py" = ["{version}"]